format.h: Fix signed/unsigned warnings in format_string_checker
clang-tidy currently complains with "implicit conversion changes
signedness: 'int' to 'unsigned int" in the class format_string_checker.

This stems from the fact that we write in on_arg_id a signed integer
into arg_id which is unsigned.

A straight-forward fix is changing the signedness of arg_id and num_args
from unsigned to signed. While that reduces the number of possible
arguments by a factor of two, this is still way more than we need.

The additional check in check_arg_id() is now required due to _arg_id_
being signed.